% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTA.R
\name{gen_fig}
\alias{gen_fig}
\title{Generate figures of COTA's result}
\usage{
gen_fig(gene.pair, p.wes, eta.wgs = 1e-05, pic_dir)
}
\arguments{
\item{gene.pair}{data frame of identified gene pair, output by calc_pair().}

\item{p.wes}{a vector of length K. p value for gene 2 -> trait. Both p.trans and p.wes should contain the gene name.}

\item{eta.wgs}{the significant threshold for WES gene}

\item{pic_dir}{the directory to save figures}

\item{sig_gene2}{gene 2 identified by COTA which are also WES significant genes, output by calc_pair().}

\item{non_sig.gene2}{gene 2 identified by COTA which are NOT WES significant genes, output by calc_pair()}

\item{conservation.score}{data.frame including the conservation scores which should contain columns: 1. 'pLI.score', 2. 'EDS', 3. 'RVIS', 4. 'GeneSymbol'}
}
\value{
a list includes identified pairs, combined pairs (loci based), identified core genes (also WES significant), and identified core genes (NOT WES significant)
}
\description{
Generate analysis plots related to COTA results
}
