% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTA.R
\name{med_gene}
\alias{med_gene}
\title{Apply COTA to identify candidate core genes.}
\usage{
med_gene(p.trans, p.wgs, ref.table, candidate, target.fdr = 0.1, dist = 5e+06)
}
\arguments{
\item{p.trans}{p value for trans genes. M*N dim matrix. There are M rows for gene 2. N columns for gene 1. Each entry is the p-value of gene 1 -> gene 2.}

\item{p.wgs}{a vector of length K. p value for gene 2 -> trait. Both p.trans and p.wgs should contain the gene name.}

\item{ref.table}{reference data for position information.}

\item{candidate}{vector of candidate gene name (for gene 1) which should be the subset of colnames of p.trans}

\item{target.fdr}{the pre-defined FDR level in the BH step, the default value is 0.1.}

\item{dist}{the distant between the gene and trans-genes, the default value is 5e6}
}
\value{
A list of candidate genes 1, identified pair matrix, and COTA's p-values
}
\description{
Apply COTA using p-values from TWAS and eQTLGen data to generate COTA's p-values and identified pairs.
}
